requirements.txt
torch
datasets==2.18.0
tokenizers
tqdm
matplotlib==3.8.0
numpy==1.26.4
pillow
PyYAML
#需要使用GPU运行时，请使用该指令(注意自身环境pytorch版本的对应)：conda install pytorch==1.12.0 torchvision==0.13.0 torchaudio==0.12.0 cudatoolkit=11.3 -c pytorch